flask/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.4
defusedxml==0.7.1
Flask==3.0.2
Flask-Login==0.6.3
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
paramiko==3.4.0
psycopg==3.1.18
psycopg-binary==3.1.18
pycparser==2.21
PyNaCl==1.5.0
python-gvm==24.1.0
typing_extensions==4.9.0
Werkzeug==3.0.1
